api_server/requirements.txt
attrs==20.3.0
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
cryptography==3.2
Flask==1.1.2
Flask-AuditLog==1.0
Flask-Cors==3.0.8
Flask-No-Cache==1.0
Flask-SSLify==0.1.5
Flask-Testing==0.7.1
google-api-core==1.20.0
google-auth==1.17.1
google-cloud-core==1.3.0
google-cloud-datastore==1.12.0
google-cloud-firestore==1.7.0
google-cloud-kms==1.4.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.33.2
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
jwkaas==0.7
MarkupSafe==1.1.1
numpy==1.19.4
openapi-spec-validator==0.2.9
pandas==1.1.4
pkg-resources==0.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
swagger-ui-bundle==0.0.6
urllib3==1.25.11
Werkzeug==1.0.1
XlsxWriter==1.3.7
zipp==3.4.0
